mapper/requirements.txt
Flask==2.2.3
Flask_RESTful==0.3.9
jmespath==1.0.1
numpy==1.22.4
paho_mqtt==1.6.1
prettytable==3.7.0
python_dateutil==2.8.2
pytz==2023.3
scikit_learn==1.2.2
tensorflow==2.12.0
waitress==2.1.2
